Completed #56 so it can be done with a timestamp. #37 was already working but now we can use it. #46 guessing phase always closed automatically but that is another issue closed.

+++ src/test/java/ch/uzh/ifi/hase/soprafs21/rest/mapper/ScoreboardDTOMapperTest.java
@@ -26,7 +26,7 @@ public class ScoreboardDTOMapperTest {
         players.add("Anton");
         players.add("Mike");
         scoreBoard.setPlayers(players);
-        long [] newScore = {545654, 456465};
+        int [] newScore = {545654, 456465};
         scoreBoard.setScore(newScore);
         int[] ranking = {1,2};
         scoreBoard.setRanking(ranking);
